Split data classes into their own source files
